create vue wallet services plugin package
