+++ packages/composables/wallet-services-plugin-vue-composables/package.json
@@ -1,0 +1,54 @@
+{
+  "name": "@web3auth/wallet-services-plugin-vue-composables",
+  "version": "9.0.0-alpha.6",
+  "description": "Vue composables to simplify integration with web3Auth wallet services plugin SDK.",
+  "keywords": [
+    "web3Auth/wallet-services-plugin-vue-composables",
+    "web3Auth",
+    "blockchain",
+    "vue composables",
+    "web3Auth composables",
+    "webAuth wallet-services plugin vue composables"
+  ],
+  "sideEffects": false,
+  "main": "dist/lib.cjs/index.js",
+  "module": "dist/lib.esm/index.js",
+  "unpkg": "dist/walletServicesPluginVueComposables.umd.min.js",
+  "jsdelivr": "dist/walletServicesPluginVueComposables.umd.min.js",
+  "types": "dist/types/index.d.ts",
+  "author": "Torus Labs",
+  "homepage": "https://github.com/Web3Auth/Web3Auth/tree/master/packages/composables/wallet-services-plugin-vue-composables#readme",
+  "license": "ISC",
+  "scripts": {
+    "dev": "torus-scripts start",
+    "build": "torus-scripts build",
+    "lint": "eslint --fix 'src/**/*.ts'",
+    "prepack": "npm run build"
+  },
+  "files": [
+    "dist"
+  ],
+  "dependencies": {
+    "@web3auth/base": "^9.0.0-alpha.6",
+    "@web3auth/wallet-services-plugin": "^9.0.0-alpha.6",
+    "@toruslabs/base-controllers": "^6.1.2"
+  },
+  "peerDependencies": {
+    "@babel/runtime": "^7.x",
+    "vue": "^3.4.x"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Web3Auth/Web3Auth.git"
+  },
+  "bugs": {
+    "url": "https://github.com/Web3Auth/Web3Auth/issues"
+  },
+  "engines": {
+    "node": ">=18.x",
+    "npm": ">=9.x"
+  }
+}
