+++ packages/composables/wallet-services-plugin-vue-composables/src/composables/index.ts
@@ -1,0 +1,1 @@
+export { useWalletServicesPlugin } from "./useWalletServicesPlugin";
