+++ packages/composables/wallet-services-plugin-vue-composables/torus.config.js
@@ -1,0 +1,1 @@
+module.exports = require("../../../torus.config");
