+++ packages/composables/wallet-services-plugin-vue-composables/tsconfig.build.json
@@ -1,0 +1,6 @@
+{
+  "include": ["src"],
+  "compilerOptions": {
+    "jsx": "react-jsx"
+  }
+}
