+++ packages/composables/wallet-services-plugin-vue-composables/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../tsconfig.json",
+  "include": ["src", "test"],
+  "compilerOptions": {
+    "jsx": "react-jsx"
+  }
+}
